Split pdp VersionLoader to new file - also we re-resolve pathways data for log in version 2
